Check number and values of testee results
Add a check that there are at least as many values return from the tested form as listed in the list of expected values.  This doesn't assert equality because there are situations where you will return more than one value, but you only care about the first several (e.g. a function that ends with a call to gethash may not care about the second return value).